Service worker handles bookmarks
The SitePassword metadata is now stored in a bookmark named SitePasswordData, and its ID is stored in chrome.storage.local.

Three cases are needed.  Both the bookmark and its ID are saved, the bookmark but not the id are saved, and neither is available.  This version handles all three cases.

Passwords are not yet calculated.  Fixing that problem will involve working on the application page.