Bookmark updating properly
The previous version used window.onunload to persist any updates to the metadata.  That approach only works for background pages, which are not supported with Manifest V3.  Instead, I'm updating the metadata using window.onblur.  Fortunately, the only way to use the password is to click outside the popup window, so this approach doesn't change the user's behavior.

There are still problems.  The popup reports that a password field can't be found when it can, and the master password has to be entered on every use.  This latter problem arises because there is no longer a persistent background page to keep it.