Getting closer
Things are working with content scripts, but I have some kludges in the code.  The main issue is that I currently have bg.rs and findpw.js both running on the page as content scripts and in the popup.  They need to be refactored.

I also still need a way to preserve the master password across uses.