Basic functions work
But there's still some debugging to do.  The big one is clearing the site password in bg.js when changing pages.

At this point, I can set up the metadata for a site, retrieve it, enter the master password, and fill in the password field on the web page.

However, going to another page and coming back reports an error on the console.  The master password sticks around for a while in the background script, but I still need a longer term place to stash it.  I've also got a ton of console.log statements that need to be removed.