Use modules to avoid duplicated code
I need both the popup and the background page to be able to compute passwords, so I split out the calculations into a module generate.js.  That change required that I use modules so I can import generate from both places.

There is still some debugging to be done, but the basic function of filling in the password field works.