Master password preserved if it's available
The service worker bg.js holds the master password while it is active.  This version sends the master password back to the popup when it opens so the user doesn't have to enter it again.

Unlike the Manifest 2 version, the service worker in Manifest 3 will close after a while, so this approach is not sufficient to get to the ease of use of the earlier version.