Correctly filling in password without opening popup
bg.rs has to compute the password if the popup is closed.  It can do that if it knows the master password, sitename, and username for the page.  bg.rs is the only component that tells the content script the password for the site.

The password is only calculated in ssp.js so it can be copied to the clipboard if the password field can't be found.  It also calculates it while the user is typing, which isn't necessary but is cool.