Fix problem with multiple events
When the cursor is on the master password field in the popup, and the user clicks in the popup window, the password field's onblur handler gets invoked.  If the user clicks outside the popup window, the only the popup's onblur handler gets called.  I was only sending the master password to bg.js in the first case.

This version uses a better approach, sending data in the popup's onmouseleave event.  Since any of the fields could have been changed, I send the domainname, mastepw, sitename, and username.