Solved "message port closed" problem
The content page was getting undefined when asking the background page for its metadata.  The problem was a silent "message port closed" error.  I had to check chrome.runtime.lastError to find out that it was happening.

The usual fix is to return 'true' from a message listener that uses sendResponse().  That was only part of the problem.  It turns out that you get the same error if you directly use an async function in the listener.  It doesn't matter if you use await or .then(); you get the message port closed problem.

The solution is the first answer to https://stackoverflow.com/questions/44056271/chrome-runtime-onmessage-response-with-async-await, create an async function that does the await.  (That's gotta be a chrome bug.)  Oh, and don't forget to return 'true' from the listener.