Popup async working
I had a hard time getting async working in the popup.  The popup would ask the service worker for metadata and return before getting it.  The trick was sending a closure to the function calling sendMessage that did the local updates that depend on the metadata.

Still some debugging to do, but at least this part is working.