Basic functions working
I was getting the strangest behavior.  The popup would send a message with certain values for sitename and username in bg.settings, and bg would get the message with different values for those variables.  I was (and am) totally baffled.  However, I changed some other stuff, and the problem went away.

This version properly computes the site password, fills in the password field on the web page, and persists the changes.  Unfortunately, it does not properly get the proper settings to the popup on the next try.