Fixed a minor usability glitch
I put a placeholder message in the password field on the web page.  It says "Click Here for Password" if the master password, site name, and user name fields are filled in.  Otherwise, it says "Click SitePassword."  I was changing to the first message while the user typed into the user name field.  This version changes the message only when the user leaves that field.