Only update placeholder when mouse leaves field
The page's password field gets a placeholder that either says "Click here" or "Click SitePassword" depending on whether the master password, sitename, and username fields are all populated or not.  I was changing the placeholder on key up, but that's misleading.  I now change only on blur, e.g., when the mouse leaves the field, which is likely when the field has the correct value.