Basic functions working
I can fill in the pop menu, calculate the site password, switch to another page, come back, click the web page password field, and get the correct password without opening the popup.

Now I need a way to persist the master password.  I can only compute the password after returning to the page if the service worker is still active.  Unfortunately, it closes after a couple of minutes (supposes to be 5, but I think it's less).