Create a new tab to hold the master password
Manifest 3 extensions do not have a persistent background page, which means I have no place to hold on to the master password.  I'm going to try opening a new tab and putting the master password in its session storage.

This commit has code to create the tab only if it does not yet exist.  There is no code yet to communicate with any of the other components.  One complicating factor is that this new tab can't receive messages from the other components; it can only receive them and return a result.

Be prepared to abandon this approach.