Mostly working but masterpw is showing up in the bookmark
A key selling point of SitePassword is that the master password is never stored; it exists solely in the memory of a running process.  However, by some mysterious mechanism, the master password appears in the metadata that does get stored, and I can't find out how that's happening.  I could just remove the value before storing it, but the master password isn't supposed to be there in the first place.