Got rid of bg.settings.masterpw
I thought the problem was in bg.js, but it was really in ssp.js.  I have a generic function for handleblur on settings.  Unfortunately, I was using that function for masterpw, so it happily set bg.settings.masterpw.  No wonder I couldn't find it by looking for "bg.settings =".