Remove redundant element in the response to getMetadata
The response to getMetadata included bg, hpSPG, and masterpw.  The last of these isn't necessary, since bg.mastepw has the same value.

I also forgot to remove a debugging change that added the string "bar" to the master password on keyup.  I didn't see the error because the correct value was used onblur.