Some experiments with session storage
I had the wrong mental model of what constitutes a session for an extension.  For the popup, I thought it ended when the window closed.  For the service worker, I thought it was when the service worker disconnected.  Neither is true.  The session appears to be the duration that the extension is active, e.g., until the browser is closed or the extension is explicitly disabled.

That being the case, I no longer need this branch.  I can go back to running bg.js in the service worker.

I still have many questions.  Do the service worker and popup share the same session storage?  Is an extension's session storage safe for the master password?