Fixed a bug that put masterpw into bg.settings
bg.settings is the metadata that is saved in the ssp bookmark.  It must not include the user's masterpw, but without this fix it does.

The problem is in handleblur, a generic method that updates bg.settings whenever a field gets modified.  One of those fields is masterpw.  This commit tests for the field and only adds other fields to the settings