Use chrome.storage.session to hold masterpw
The "session" in chrome.storage.session refers to the browser session.  That means anything that's stored in it survives the popup closing and the service worker detaching.  That's exactly the behavior I need to keep the user's master password in a safe place.  It's safe because session storage does not appear in the file system.