Works if inspecting popup but not otherwise
The difference is how a "The message port closed before a response was received" error.  When not inspecting the popup, the window closes before the siteData message can be delivered to the service worker.  The window stays open when the inspector is open, so the message gets delivered.

I'm doing this commit because I have a bunch of changes not related to this problem that I don't want to lose.