Get correct messages in site's password field
There are 3 things the user might need to do to get the password into the site's password field.  If the master password, site name, or user name aren't set, the user must click on SitePassword.  If those values are set, the user can click on the first found password field.  If there's more than one, such as confirming your password when changing it, the user can paste the password into subsequent password fields.  This commit sets the proper instructions as the placeholder of the site's password field.