Copy to clipboard when bg.js calculates password
I had forgotten that bg.js calculates the password when the masterpw, site name, and user name are all in the metadata.  That means bg.js must copy the site password to the clipboard, but it doesn't have access to the DOM.  Fortunately, I can do that in findpw.js when it gets the site password from bg.js.