Explain why popup needs all the saved metadata
At first glance, it appears that the popup does not need all the metadata stored in the bookmark, just the settings included in bg.  However, the popup does need all that data for both the phishing check and to export the site data.  If it was just the latter, I could transfer all the data from bg.js on the rare occasions it is needed.  However, the phishing check is done all the time.

This is not the first time I've forgotten this fact, so I'm adding a comment to remind me.  Now all that needs to happen is for me to actually see it.