Fixed a problem with placeholder
findpw.js looks at all the input fields on the web page and collects those that are password fields into an array.  I was doing the counting on every message in addition to when the page loaded, and I wasn't initializing the array.  The result was that each call added the same password field to the array.  The result was that the placeholder was set as if the password field wasn't the first one in the list.

In case you're wondering why I don't just do the count when the page loads, some pages might add password fields later or replace the one I saw before.