Fixed problem with placeholder
I wasn't updating the placeholder when ssp.js calculated the site password.  The fix was one line, but I noticed my copy-to-clipboard test was wrong, so I fixed that, too.  I also put back in the default focus to save the user a click.

It takes several steps to recover from a bug that messes up the metadata, so I decided to save a snapshot of a good version.  Now I can just reset to this good value from the console.