Cut out a line I shouldn't have
When findpw.js gets a password from either bg.js or ssp.js, it is supposed to copy it to the clipboard if there's not exactly one password field.  I had inadvertently deleted that line before the last commit.

My test didn't catch it, because the password was on the clipboard from an earlier test.  The trick is to remember to put garbage on the clipboard before clicking on the first password field when there is more than one.