Fixed a long standing problem
Sometimes after setting the value of a field, that value disappears after the focus leaves the field.  I think that happens when the field is created dynamically, but I'm not sure.

At any rate, my solution with Version 0.98 was to manually type and erase a character in the field before changing focus.  I can't do that from JavaScript, so I'm doing the next best thing.  This commit tests to see if I'm in the bad case.  If so, it puts the text on the clipboard and changes the placeholder message to pasteHere.

NB: This change does not work on all web sites.

NB: There's a UX problem in that the user clicks on the field, but the only thing that happens is that the placeholder changes.