Set onclick for password field where I find the field
I was setting the onclick function for the password field in the window.onload function.  That didn't catch the times when the password field was generated later, such as when the credentials were invalid and had to be reentered.  This commit sets the onclick function where I find the password field, so it sets it for all of them.