Fixed bug with legacy
In the Manifest 2 version, the popup had direct access to variables on the background page.  That's not true with Manifest 3, but I had a vestige of the old version.  The fix involved moving "legacy" from the top level of bg to bg.settings, which is a more natural place for it anyway.