Don't send site password if username is ""
I send the site password to the content script if I don't find any password fields, so findpw.js can put the password on the clipboard.  However, most of the time, that's because the page really doesn't have any password fields, or it's a site that hasn't been set up for SitePassword.  I can tell that's the case by testing to see if the username has been set.