Works with password field in an iframe
On at least one site.  This commit resolves a number of problems.

I was taking the domain name from the content script, but it might have been compromised to spoof it.  There was also a problem identifying the domain name if the window.onload was an iframe, which has a different URL.  Both problems were fixed by extracting the domain name from sender.tab.url.

A second problem was finding the password field if it is in an iframe.  This problem was fixed by setting "all_frames" to true in the manifest.  Every iframe now looks for any password fields it may have.  Because the code is async, I can't save the found fields across invocations.  I have to find them again every time findpw.js needs them.

This commit also finds the password fields with a document.querySelect cribbed from Bitwarden.  I don't know if it's better than just using document.getElementsByTagName("input"), but they must have had a reason.