Fixed case when page changes placeholder
I was testing to see if the password field placeholder was "clickHere" in case the password wasn't ready and it said "clickSitePassword" instead.  However, some pages change the placeholder before or on the user's click.  I now check to see if the placeholder is any of the options other than clickHere.