Fixed URL to use for domainname
I was using the tab's domainname for calculating the site password, but that allows an iframe from a different domain to learn the password of another domain.  Instead, I need to use the origin of the message sender.  In a future commit I'll fix up ssp.js to do the same.