Fixed problem with password field in iframe
There can be many iframes on a page that don't contain a password field and one or more that do.  I was having two problems.

First, the iframes load concurrently, so setting up the userid and password fields in an iframe having these fields would be overwritten on some runs by iframes that did not.

Second, the popup was always showing the domain name of the tab, not the iframe with the selected password field.

I fixed both these problems by introducing a map from tab domain name to iframe domain name.  Now bg.js remembers the iframe domain name and can tell the popup the value.