Two minor changes
I was sending page info to the service worker even if there are no password fields found on the page.  That's not needed because the user will have to open the popup, which has the same effect.

This commit also sets the focus on the user id field if there's no user id and on the password field if there is.  A small change, but it saves the user a click.