Added visibility test for userid field
It doesn't make sense to check for visibility of password fields.  You're not going to click on what you can't see.  However, it does make sense to check the visibility of fields when looking for the userid fields, since there might be hidden fields between the password and userid fields.

I also added a visibility test condition.  I don't believe it should be necessary, since there are ways to say a field is hidden outside of the aria attributes, but that's what's used on the ebay signup page.