Two substantive changes
There is a click jacking attack if the password gets filled in when a user clicks on an invisible password field.  This version sets the onClick callback only for visible fields.  Unfortunately, visibility testing is hard, so the code sometimes gets it wrong.  The only downside for the user is that sometimes a paste instead of a click is what's needed.  An alternative would be to paste all the time, but that choice always puts the password on the clipboard.

The second change is to remove the password from the clipboard after 10 seconds.  There's still a chance that some other application will get it, but the window is small.  If the user doesn't paste in time, the place holder changes to tell them to open the popup.  When they do, they get another 10 seconds to paste.

Since there's no way for JavaScript to clear the clipboard, I paste "".  Something better is needed on systems that keep multiple elements on the clipboard.