10 seconds isn't long enough
Clearing the password from the clipboard after 10 seconds is too fast.  It takes a while for the user to find the password field on some pages.  The 10 seconds often expires while the user is interpreting the placeholder.  Better to make the window of opportunity to get the password off the clipboard longer, currently 30 seconds, than to confuse users.