Fixed problems with multiple password fields
I decided that I should only deal with the first two password fields on a page.  The second is always(?) used to confirm the password on signup.  Other fields with type = password are other thing, such as the answers to security questions.  The code needed some changes to implement this decision.