Labels on the userid and password fields make fields look messy
Labels on the userid field sometimes obscure the userid.  Those on the password field often obscure the placeholder.  The result is ugly.

I spent the day trying to hide or get rid of the labels, but what works on one site fails miserably on another.  I'm leaving my tests in place in case I come up with a solution even though I do an immediate return.