Fix previous commit
The previous commit looked for password fields in the mutation observer.  Unfortunately, I wasn't sending the info it found to the service worker.  That broke any page that modified the page to add a password field.  Note that I got into a loop without the test on the call to sendpageinfo().