Fixed popup to deal with multiple password fields
I had changed findpw.ps to handle multiple password fields differently, but I hadn't made the corresponding changes to the popup.

When there are no visible password fields, the popup puts the site password on the clipboard and clears it after a minute.  That's longer than the 30 second when the content script puts the site password on the clipboard to give the user extra time if the popup has to be opened.