Add tooltip to password fields
Some web sites don't display the placeholder for reasons I can't figure out.  Others have labels or even default text that makes the placeholder illegible or completely hides it.  Adding a title attribute to the field gives me a tooltip with what's in the placeholder I want the user to see.