Make it easy to turn off label hiding
Some sites put labels in the userid and password fields that overlay the userid and placeholder SitePassword sets.  It's not a problem functionally, but it is ugly.

Hiding labels is possible, but it can be dangerous by making the password field inaccessible.  I haven't encountered any pages that have this problem, but I'm leaving my options open.

Right now, the boolean controlling whether or not labels are hidden is hardwired in findpw.js.  However, I'm considering making it part of the metadata associated with a site.