Got rid of most of the annoying labels
Web sites often use labels on input fields to tell the user what goes there.  Sometimes these labels are inside the input field, which looks ugly when SitePassword has filled in the userid and put a placeholder in the password field.  This commit fixes all but 11 of the 44 sites that need the labels hidden.  Labels don't overlap the input fields on 41 sites and are left visible.  Only 14 sites do not use labels.  One site has a blank login page.