Added additional search for labels
Labels are supposed to have a "for" attribute that says what element they are labeling, but not all sites do that.  Some sites put the "for" in a span inside the label.  This commit checks for both.  Of course, that means it finds labels that use "for" twice, which is why I dedup the resulting array.

Even though I haven't encountered the problem, I now check the visibility of the labelled element after I've hidden the label and undo the change if the element became invisible.