Remove implicit copy to clipboard
Version 0.98 found password fields on only about 60% of the sites, so it made sense to copy the site password to the clipboard without any user interaction.  Version 0.99 almost always finds the password field, so that decision no longer makes  sense.

This commit asks the user to click on the site password in the popup to copy it to the clipboard.