Fix mutationObserver problem
I need to update the userid and password fields inside the mutationObserver callback in case the page changed what I set.  Unfortunately, observer.disconnect() doesn't appear to work inside the observer callback.  The result was continually hitting the callback because of changes I made.

This commit bypasses the update if the field already contains the value I had set.