Put mutationMax back
signin.ebay.com changes the placeholder to "" every time I set it.  It also sets the label's class to the value it already had even though I only set the label visibility once.  The result is an infinite recursion of the mutation observer  that I can't fix.  It also explains why the placeholder is missing; it gets reset after I've done mutationMax iterations.  I've set mutationMax to be big enough not to miss any intended updates but not so large as to make the user wait for a page load to finish.