Two improvements
Fixed the slow load on myacm.acm.org.  Using window.onload waits for all resources to be loaded, which is very slow (30 seconds) because it times out fetching some of its own resources.  Using document.onload is faster because it doesn't wait.  However, you have to check the ready state in case the document is already loaded by the time you hit that statement.

The second improvement is dealing with cycles in the mutation observer.  For example, signing.ebay.com undoes my placeholder change, which results in a mutation event in which I change the placeholder.  Lather, rinse, repeat.

I was stopping after a fixed number of tries, but how big should that number be?  In this version I try a small number of times (10), but I try another 10 after a short wait (0.5 seconds).  Now it's faster than cycling 500 times, but I catch most mutations.