Forget a domain name
There may be times when you want to forget the settings for a domainname, e.g., you accidentally set up for a malicious site.  You now have that option.

Forgetting a domainname does not delete the settings, it just disassociates the domain name from them.  That way you can undo the forgetting by just setting the sitename to the value you used before.