Completely forget forgotten domainname
The previous commit left the sitenames entry even if the deleted domain name was the last reference to this sitename.  The result was trying to use the sitename again resulted in a phishing warning.  That warning would probably confuse the user.

This commit removes the entry from sitenames if the forgotten domain name is the last reference to that site name.