Find password field in shadow root
Some (a very few) sites put the password field in a shadow root.  A normal search won't find it.  I have to walk the DOM to find the shadow root then search inside of that, which only works if the shadow root is Open.

The downside is that that I have to walk the DOM on every page, not just those with a password field, and every time that page changes in case the change added a password field.  Informal measurements show little if any performance difference, but I can imagine complex pages where the difference will be noticeable.