Some code clean up
The original code assumed one password field.  If there was more, as a confirm password field on a create account page, the user would paste the password into the second field.  I reduced the use of the clipboard by allowing the user to click on the second field to insert the password.  The problem is that the code structure didn't reflect that change.

The variable cpi (for count, pwfield, idfield, but I removed count) now has an array of password fields instead of having a single one and remembering the second field in another variable.