Turned off search for shadow root
Finding a password field in a shadow root requires searching the DOM on all pages every time the page changes.

Preliminary tests show that's not a performance problem, but I'm still worried that it might be on some sites.  The downside of not searching is needing to open the popup on pages where the password field is inside a shadow root.  Of my 100 test sites, only cvs.com does that.

The search can be turned back on by commenting out the one line I changed in this commit.