Hide persona field in popup
I haven't tested multiple personas, so I think it's best to hide that option from the user until I do.  The main benefit of multiple personas is a different master password for each one, but you can get that by just changing the master password.  You don't get the other benefit, which is different settings for the same web site.