Fix typing in password field
There may be a time when you want to type a password instead of using the provided one.  However, some sites kept refilling the field with the provided password.  This commit takes care of that problem by detecting if the update to the field is due to a key down event.

It's odd that keydown fires when the field is updated programmatically, which means no key is pressed.  The way to distinguish the two ways to update the field is to look for event.key.  It's undefined if the update is programmatic.