Remove "legacy" from code
The intern chosen to implement the HP Antiphising Toolbar made some usability mistakes, such as making the site name case sensitive and not trimming white space.  I set up SitePassword with a legacy option so it would generate the same passwords.

I've decided that it's easier to just change the passwords on the few sites that are still using legacy passwords.  Hence, there's no longer a need for "legacy" in the current version.