Removed masterpw from console log
It's important to keep a master password secret, so printing it on the console seems risky.  Still, it's useful for debugging to know if it has been set or not.  I defined a function isMastePw() in generate.js because it's used by both bg.js and ssp.js.  The return value says if it's defined or not as a string that can be included in a console.log entry.