Set up to work with browser version
I've set up a web site that does the same password calculate as the extension, so you can calculate your passwords when you're on a device that doesn't have the extension installed.  Unfortunately, that page doesn't have access to the bookmarks that hold your metadata.

This change stores a bookmark for each domain name with its metadata.  The web page will be changed so you can paste the bookmark, and the page will extract the metadata.