Turned shadow root search back on
I originally turned it off because I was afraid it was generating too much overhead, but that doesn't appear to be the case.  Even on a site like mlb.com that updates the page every second or so the CPU utilization, which is highly variable, doesn't appear to be even 1% higher with the search on.