Fixed popup message
The popup was sometimes reporting that there were no password fields on the page even though they were.  The problem was that the popup was asking the service worker for the information.  The service worker had gotten that data from the content script, but it forgot it when the service worker went idle.

The fix is to have the popup ask the content script for the info.