Explain a "clever" hack
SitePassword uses bookmarks for two purposes.  One is to hold the pieces of the stringified hpSPG.  These bookmarks have titles that are string representations of integers, "0" for example.  The other bookmarks hold the settings for a particular domain name so that users can paste them into the web version.  The title of each of these is the domain name.

The hack is to distinguish the two by the fact that the former can be converted to integers, and the latter cannot.  That's obscure enough to warrant a comment.