Fix an annoying UX glitch
I want to close the popup on mouse out, but that results in lost messages since the sending end of the pipe gets closed before the service worker gets the message.  The result of not closing the popup is that the user needs two clicks to fill in the password field.  The first closes the popup, and the second fills the field.  Delaying closing the window fixes this problem.