Update license agreement
The original version was done while I was working for HP.  HP required a particular license agreement in each file.  I have been unable to reach the legal department at either HP Inc. or HP Enterprise to ask how to update it.  This commit is what I came up with.