Took out Forget option
Forgetting a domain name is harder than I thought it was going to be.  I can easily remove it from hpSPG, but I can't figure out how to delete the bookmark with a title of the domain name.  Bugs in it keep messing up the bookmarks, and I haven't been able to figure out why.  I think I'll leave this feature for a future release.