Fixed race condition
findpw.js sends the SiteData message twice, once for document.onload and once for window.onload.  The latter can arrive before processing the first message finishes creating the bookmarks folder.  The result is two of them.  My fix is to use a flag to make sure the bookmarks folder is created only once.