Fix handling of iframes
The origin of an iframe is different from that of the containing page, which is the domain name used by the popup.  Things work properly if I set up the metadata for the iframe's origin, but the popup fields don't get populated because there's no metadata for that domain name.

You might think users would never open the popup once they'd set up a site, but they will when they need to change a password.  Not seeing their site and user names would be confusing.

I fixed the problem by adding an entry to persona.sites for each ancestor of the iframe.  To do that, I have to send the ancestor list from the content script.