Fixed spurious popup message
The popup was usually, but not always, reporting that no password fields were found on sites with the password field in an iframe.  The problem is that each iframe sends the count of password fields separately to the popup, which was reporting the correct message for the last message it received.

This commit remembers if any of the messages reported finding a password field and sets the message appropriately.