Fixed popup problem
When you open the popup on a site that has already been set up, you expect to see your site and user names filled in.  It's necessary when you want to update one of them, but it wasn't happening on sites that put the password field in an iframe.  That's because the popup used the domain name of the page, not that of the iframe.

The fix is to have the content script remember the domain name and password count of an iframe that contains password fields and send that to the popup on request.

I had to change the order of operations to make this work.  The popup was asking the service worker for the settings before asking the content script for the password count.  i've reversed that order, so the popup knows what domain name to use when asking for the settings.