Fix erroneous bookmark
I need to distinguish between the domain name of a frame containing a password field and the domain name of the page containing the iframe.  The settings are the same for all pages that have the same iframe, but the name displayed should be that of the page, not the frame.