Fix clearing site password from clipboard
Your site password is put on the clipboard when you click on that field in the popup.  Since it's not a good idea to leave security critical information on the clipboard, I was trying to replace it with "" after some number of seconds.  Unfortunately, I was closing the popup, which caused the writeText in it to fail with a "Document not focused" error.

This version has the content script clear the site password, since it comes into focus as soon as you mouse out of the popup.

There are a couple of problems with this approach.  First, the write of "" will fail if the user hasn't moused out of the popup before the timer expires, which I think is highly unlikely.  The second is that the user might put something else on the clipboard before the timer expires, which will be replaced when it does.  This case is also highly unlikely if the timeout is short, say 10 seconds.

A better solution would be to check if the clipboard has the site password before clearing, but the content script doesn't have permission to read the clipboard for very good security reasons.