Stop clearing clipboard
It's not a good idea to leave a password on the clipboard for very long.  The problem is knowing how long is long enough.  There are two problems.

If the time is too long, the user might have put something else on the clipboard.  You could check to see if the password is still there, but reading the clipboard from Javascript is tightly controlled.

If the time is too short, then the user might not have finished pasting the password.  (People do get distracted.)  Then you need a way to tell the user to paste it again, but it's not obvious how to make that clear.

The strategy used here is to tell the user to put something else on the clipboard when done pasting.