Deal with multiple SitePasswordData bookmarks
Synchronizing between machines is tricky, especially when they aren't all online when changes are made.  Google does the sync for me, but I saw a case where it duplicated the SitePasswordData bookmark folder.

That shouldn't happen, but to be safe, I'm having the content script put up an alert telling the user to rename or delete all but one.