Don't keep creating mutation observers and listeners
Long ago in a galaxy far away the code in startup() was executed only once.  That's no longer the case; it's called repeatedly.  Without these changes, I was creating a new mutation observer and listener on every call.  That's not necessary and it could lead to a loss of events.