Removed delete database entry
In manifest 2, the database variable was shared between the background page and the popup.  As a result, I could remove an entry from the database in the popup if the user cleared the site name field.  That doesn't work with manifest 3, so I took it out.

Along the way, I changed some variables from var to const.