Fixed popup anomalies
The popup was showing mistaken information, such as the settings for the last time I opened the popup on another site.  It would also say no password fields were found when they were.

The first problem was due to storing the last used domain name and password field count in local storage and retrieving that information without regard to which page was asking.  I replaced that with simply doing the password field search again and not using local storage.

The second problem was two additional places where I got database.domains and database.sites mixed up.