Fixed mapping error
When the password field is in an iframe, the origin of the request is that of the iframe, not the domain name saved for that site.  I introduced pwfielddomainname to hold that mapping, but I switched the key and value so the mapping went the wrong way.

Fixing that error means I no longer need settings.displayname, which I introduced to work around what I now know was a mistake.