Make extension calculate same passwords as Dales web page
Dale made some changes to the password calculation that improve security and simplify the code.  Increasing miniter makes guessing harder, and adding '\t' between the inputs to the calculation avoids some accidental collisions, such as site name "foo" and user name "bar" producing the same password as "foob" and "ar."  He also simplified and improved the calculation of the set of characters used for the site password.