Simpler way to handle special characters
Including some special characters (I'm looking at you, #.) when producing the download site data file causes problems.  I didn't know it, but there is a built in function for dealing with them, which I'm now using.

I tried using base64, which is supposed to deal with the problem, but I couldn't make it work.  The chrome.downloads.download() produced no output and not error message.